Reestructuración de las vistas de Tickets
