+++ DirCom/applications/tickets/views/__init__.py
@@ -1,0 +1,2 @@
+from .export_tickets import *
+from .manage_tickets import *
